+++ R/zzz.R
@@ -43,6 +43,9 @@ if (getRversion() >= "2.15.1") {
   register_s3_method("dplyr", "transmute", "events_tbl")
   register_s3_method("dplyr", "summarise", "events_tbl")
   register_s3_method("dplyr", "as_data_frame", "eeg_lst")
+  
+  register_s3_method("dplyr", "pull", "eeg_lst")
+  
   register_s3_method("ggplot2", "ggplot", "eeg_lst")
   register_s3_method("ggplot2", "ggplot_add", "layer_events")
   register_s3_method("data.table", "as.data.table", "eeg_lst")